scripts/autopost/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
oauthlib==3.2.2
Pillow==9.0.0
pydash==7.0.6
requests==2.31.0
requests-oauthlib==1.3.1
tweepy==4.14.0
typing_extensions==4.8.0
urllib3==2.0.5